Move Rate Limits card from Overview tab to Discovery tab
Rate limits configure the same external-agent traffic surface that
the Discovery tab's crawler allow-list governs — "who gets in" and
"how fast they can go" are one coherent subject in the merchant's
mental model. Keeping them on separate tabs forced a context switch
between two decisions that belong together.

Moved in full:
  - RATE_LIMIT_PRESETS constant + getActivePreset() helper
  - Local rate-limit state (rpm, customOverride, activePreset)
  - The Card with RadioControl preset picker, conditional custom
    RPM TextControl, and Save button
  - RadioControl + TextControl imports from @wordpress/components
  - useState import addition on the Discovery side

Overview tab now owns:
  - Status banner + stats + per-agent revenue breakdown
  - Product visibility summary
  - Disable button

Discovery tab now owns:
  - Endpoint reachability table
  - AI Crawler Access (allow-list)
  - Rate Limits (placed directly below the allow-list so the
    narrative order reads "who's allowed → how fast they can go")

Backing data model is unchanged: the same rate_limit_rpm setting,
the same preset values (10/25/100), the same Save flow. This is
purely an IA move.

Quality gates: 374 PHPUnit / 1039 assertions, 42 Jest tests,
PHPCS clean, PHPStan clean, ESLint clean, Webpack build clean.

Ships queued on main, no tag — see release-cadence discussion.
